mcp_network_manager/pyproject.toml
[project]
name = "mcp-network-manager"
version = "0.1.0"
description = "An MCP server for managing network devices using Netmiko"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Network Automation Team" }]
keywords = ["mcp", "network", "automation", "netmiko", "cisco", "juniper"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "anyio>=4.5", 
    "click>=8.1.0", 
    "mcp",
    "netmiko>=4.2.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "bcrypt>=4.0.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0"
]

[project.scripts]
mcp-network-manager = "mcp_network_manager.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_network_manager"]

[tool.pyright]
include = ["mcp_network_manager"]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.uv]
dev-dependencies = ["pyright>=1.1.378", "pytest>=8.3.3", "ruff>=0.6.9"] 